requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.3.0
django-debug-toolbar==4.4.6
django-import-export==4.3.3
django-simple-history==3.7.0
Pillow==10.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.32.3
python-dateutil==2.9.0.post0
sentry-sdk==1.40.6 